Factor InterpolationTarget test work
